requirements.txt
# Core Framework
langgraph>=0.2.0
langchain>=0.3.0
langchain-core>=0.3.0
langsmith>=0.1.0

# LLM Providers
google-generativeai>=0.8.0

# Web Framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.9.0
pydantic-settings>=2.5.0

# UI
streamlit>=1.39.0

# Storage & State
sqlalchemy>=2.0.35

# Utilities
python-dotenv>=1.0.1
tenacity>=9.0.0
requests>=2.32.0
httpx>=0.27.0
rich>=13.9.0

# Logging & Monitoring
structlog>=24.4.0

# Testing
pytest>=8.3.0
pytest-cov>=5.0.0
pytest-asyncio>=0.24.0
pytest-mock>=3.14.0
httpx>=0.27.0  # for testing async

# Development
black>=24.10.0
ruff>=0.7.0
mypy>=1.13.0

